Add internal util packages at a non-internal path
KubeadmControlPlane has an internal package which contains some util packages
for machines and failureDomains. Etcd cluster controller can reuse them, but those
functions can't be used directly as they are in an internal package which can't be
imported. This commit copies over the required functions at a path from which
we can import them